+++ src/__main__.py
@@ -1,5 +1,5 @@
-from Game import Game
+from src.game import Game
 
 if __name__ == "__main__":
-	game = Game()
-	game.on_execute()
+    game = Game()
+    game.on_execute()
