+++ src/assets/food.png
