+++ src/synapse.py
@@ -1,0 +1,5 @@
+class Synapse:
+    def __init__(self, source_neuron, destination_neuron, weight):
+        self.source_neuron = source_neuron
+        self.destination_neuron = destination_neuron
+        self.weight = weight
